Bugs fixed,added similar and recommendations films into film list

+++ app/src/main/java/com/alexanderivanets/filmie/MVPAttempt/searchfilm/Result.java
@@ -48,7 +48,7 @@ public class Result {
     private Boolean video;
     @SerializedName("vote_average")
     @Expose
-    private Integer voteAverage;
+    private Double voteAverage;
 
     public String getPosterPath() {
         return posterPath;
@@ -154,11 +154,11 @@ public class Result {
         this.video = video;
     }
 
-    public Integer getVoteAverage() {
+    public Double getVoteAverage() {
         return voteAverage;
     }
 
-    public void setVoteAverage(Integer voteAverage) {
+    public void setVoteAverage(Double voteAverage) {
         this.voteAverage = voteAverage;
     }
